% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eccentric-package.R
\docType{package}
\name{eccentric-package}
\alias{eccentric}
\alias{eccentric-package}
\title{eccentric: Simulate the Central Limit Theorem Using Any Distribution}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: JP Monteagudo \email{jpmonteagudo2014@gmail.com} (\href{https://orcid.org/0009-0003-6465-6658}{ORCID}) [copyright holder]

}
\keyword{internal}
